Add support for HS105 protocol
Compared to the HS100 and HS110, where messages sent and received
begin with a four-byte header that can be safely ignored (and set to
zero in the existing code), the HS105 mini-plug requires queries to
have a correct length header matching the message length.

This changeset provides such a header when sending queries. We still
ignore the header when receiving responses from the smart plug.